Now cleans up database when adding items to cart

+++ SnackShop.Core/Services/CartService.cs
@@ -1,20 +1,18 @@
-﻿using System;
-using System.Web;
-using System.Net;
-using System.Collections.Generic;
-using System.Text;
-using SnackShop.Core.Models;
+﻿using SnackShop.Core.Models;
 using SnackShop.Core.Repositories;
+using System.Linq;
 
 namespace SnackShop.Core.Services
 {
     public class CartService
     {
         private readonly CartRepository CartRepository;
+        private readonly ProductRepository ProductRepository;
 
         public CartService(CartRepository cartRepository)
         {
             this.CartRepository = cartRepository;
+            this.ProductRepository = new ProductRepository(this.CartRepository.GetConnectionString());
         }
 
         public CartModel GetCart(string cartId)
@@ -29,11 +27,29 @@ namespace SnackShop.Core.Services
 
         public bool AddToCart(int productId, string cartId)
         {
+            // Removes any old carts still in database where cookie has expired
+            this.CartRepository.RemoveOldCarts();
+
+            // Checks if product exists before adding to cart
+            var product = this.ProductRepository.GetAll().SingleOrDefault(x => x.Id == productId);
+            
+            if (product == null)
+            {
+                return false;
+            }
+
             return this.CartRepository.AddProduct(productId, cartId);
         }
 
         public bool RemoveFromCart(int productId, string cartId)
         {
+            // Checks if product exists in cart before removing
+            var cartProduct = this.CartRepository.GetCart(cartId).SingleOrDefault(x => x.Id == productId);
+
+            if (cartProduct == null)
+            {
+                return false;
+            }
             return this.CartRepository.RemoveProduct(productId, cartId);
         }
     }
